Backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
